requirements.txt
python-docx>=0.8.11
pymupdf>=1.23.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
jinja2>=3.1.0
playwright>=1.40.0
